requirements.txt
discord.py
Flask
Pillow
gunicorn